No Jump Settling
Prevents AI jump settling, which is behavior where the AI would settle
past or in between cities founded by other players. It finds the enemy
city that is most directly between the settle location and the AI
capital, then runs a serious of checks based on that city to determine
if it's too dangerous of a settle location.